+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -7,8 +7,11 @@
 #include "generated_plugin_registrant.h"
 
 #include <native_video_editor/native_video_editor_plugin_c_api.h>
+#include <video_player_win/video_player_win_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
   NativeVideoEditorPluginCApiRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("NativeVideoEditorPluginCApi"));
+  VideoPlayerWinPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("VideoPlayerWinPluginCApi"));
 }