+++ example/windows/flutter/generated_plugins.cmake
@@ -4,6 +4,7 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   native_video_editor
+  video_player_win
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST